Fix moor feature relative movement
Previously, the location of moor feature was changed upon player
by _adding_ the player bearing changes to the feature bearings.
This is incorrect: since the feature must move in the inverse
direction to the player, the player movement must be _subtracted_
from the feature bearing.

This commit fixes that bug.