Implement command parser
The CommandParser class should be able to parse an arbitrary
sequence of _valid_ commands. The new functionality has been
incompletely validated and seems to work for all valid input,
as long as the number token is a single digit.

An exception for invalid tokens is still to be implemented, so
behaviour for invalid input is undefined.